+++ server/db/models/index.ts
@@ -1,0 +1,2 @@
+export { UserModel, type TUserDoc } from './user'
+export { RoomModel, type TRoomDoc } from './room'